Result of the port scan
This scan will attempt to:

Detect open ports on the target system.

Identify the services running on those ports and their versions.

Try to determine the operating system.

Run Nmap scripting engine (NSE) scripts to check for common vulnerabilities or configurations.

Perform a traceroute to see the path to the target.

Do all of this with a faster timing (but riskier stealth-wise) using -T4

NB: Noisy Scan: This scan is not stealthy. It will likely be logged by firewalls, IDS, or monitoring tools.

Used for Recon: Commonly used during the reconnaissance phase of penetration testing to gather as much data as possible about the target.

Requires Root: Some parts of -A (like OS detection and traceroute) may require root privileges.